OPHAKTKEH-257 Refactor interfaces
