+++ src/main/reactjs/src/components/elements/ComboBox.tsx
@@ -10,7 +10,7 @@ import {
   AutocompleteValue,
   ComboBoxOption,
   ComboBoxProps,
-} from 'interfaces/combobox';
+} from 'interfaces/components/combobox';
 
 const compareOptionLabels = (a: ComboBoxOption, b: ComboBoxOption) => {
   return a.label <= b.label ? -1 : 1;
