--- src/main/reactjs/src/interfaces/datePicker.ts
@@ -1,7 +1,0 @@
-export interface DatePickerProps {
-  value: string;
-  setValue: (value: string) => void;
-  label: string;
-  minDate: Date;
-  maxDate: Date;
-}
