+++ src/main/reactjs/src/interfaces/languagePair.ts
