+++ src/main/reactjs/src/interfaces/with.ts
@@ -1,3 +1,7 @@
+export interface WithId {
+  id: number;
+}
+
 export interface WithVersion {
   version: number;
 }
