--- src/main/reactjs/src/interfaces/withId.ts
@@ -1,3 +1,0 @@
-export interface WithId {
-  id: number;
-}
