+++ src/main/reactjs/src/redux/reducers/publicTranslator.ts
@@ -2,7 +2,7 @@ import { Reducer } from 'redux';
 
 import { APIResponseStatus } from 'enums/api';
 import { SearchFilter } from 'enums/app';
-import { LanguagePairsDict } from 'interfaces/language';
+import { LanguagePairsDict } from 'interfaces/languagePair';
 import {
   PublicTranslator,
   PublicTranslatorAction,
